+++ monolith/src/main/java/ru/gb/antonov/j710/monolith/beans/services/ProductService.java
@@ -15,6 +15,7 @@ import ru.gb.antonov.j710.monolith.beans.repositos.specifications.ProductSpecifi
 import ru.gb.antonov.j710.monolith.entities.*;
 import ru.gb.antonov.j710.monolith.entities.dtos.ProductDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static ru.gb.antonov.j710.monolith.Factory.NO_FILTERS;
@@ -79,7 +80,7 @@ public class ProductService
 //-------------- Редактирование товаров ---------------------------------
 
     @Transactional
-    public Product createProduct (String title, double price, int rest, String productCategoryName)
+    public Product createProduct (String title, BigDecimal price, int rest, String productCategoryName)
     {
         Product p = new Product();
         ProductsCategory category = productCategoryService.findByName (productCategoryName); //< бросает ResourceNotFoundException
@@ -91,7 +92,7 @@ public class ProductService
 Любой другой параметр может быть {@code null}. Равенство параметра {@code null} расценивается как
 нежелание изменять соответствующее ему свойство товара. */
     @Transactional
-    public Product updateProduct (@NotNull Long id, String title, Double price, Integer rest, String productCategoryName)
+    public Product updateProduct (@NotNull Long id, String title, BigDecimal price, Integer rest, String productCategoryName)
     {
         Product p = findById (id);
         ProductsCategory category = null;
